Fix broken Export-Package and Private-Package manifest values
The value for `Export-Package` in the jar manifest is set to
`[Ljava.lang.String;@4bb370b`, since a raw array is pused to the
`attributes` method.

This commit transforms the value to a string before calling adding it to
the manifest attributes, which will produce a valid value.

Additionally, the logic for `Private-Package` was calling a non existing
methos on the manifest object, which only worked, as there are currently
no private package instructions. It has been fixed as well.